Fix bugs of config framework. The '@INHERITED' can ref the value from another template now.

+++ src/main/java/com/github/cao/awa/lilium/config/bootstrap/LiliumNetworkConfig.java
@@ -5,14 +5,14 @@ import com.github.cao.awa.lilium.annotations.auto.config.UseConfigTemplate;
 import com.github.cao.awa.lilium.config.LiliumConfig;
 import com.github.cao.awa.lilium.config.instance.ConfigEntry;
 import com.github.cao.awa.lilium.config.metadata.MetadataConfig;
-import com.github.cao.awa.lilium.config.template.bootstarp.LiliumNetworkConfigTemplate;
+import com.github.cao.awa.lilium.config.template.server.bootstarp.network.LiliumNetworkConfigTemplate;
 
 @UseConfigTemplate(LiliumNetworkConfigTemplate.class)
 public class LiliumNetworkConfig extends LiliumConfig {
     @AutoConfig("metadata")
     public final ConfigEntry<MetadataConfig> metadata = ConfigEntry.entry();
     @AutoConfig("bind-port")
-    public final ConfigEntry<Integer> bindPort = ConfigEntry.entry();
+    public final ConfigEntry<Long> bindPort = ConfigEntry.entry();
     @AutoConfig("use-epoll")
     public final ConfigEntry<Boolean> useEpoll = ConfigEntry.entry();
 }
